requirements.txt
streamlit
opencv-python
torch
torchvision
easyocr
numpy
pandas
pillow
ultralytics